Add /team/new page
