+++ src/ui/components/Account/index.tsx
@@ -1,12 +1,10 @@
 import React, { useEffect } from "react";
-import useAuth0 from "ui/utils/useAuth0";
-import { setUserInBrowserPrefs } from "../../utils/browser";
-import Library from "../Library/index";
+import { useRouter } from "next/router";
 
-import { PrimaryLgButton } from "../shared/Button";
-import { OnboardingContentWrapper, OnboardingModalContainer } from "../shared/Onboarding/index";
-import { useRouter } from "next/dist/client/router";
-import { isTeamMemberInvite } from "ui/utils/onboarding";
+import useAuth0 from "ui/utils/useAuth0";
+
+import Library from "../Library/index";
+import Login from "../shared/Login/Login";
 
 /**
  * Create a (host relative) URL with the given parameters. Used for replacing
@@ -22,50 +20,6 @@ function createReplayURL(currentParams: URLSearchParams) {
     }
   });
   return `${path}?${newParams.toString()}`;
-}
-
-function WelcomePage() {
-  const { loginWithRedirect } = useAuth0();
-  const forceOpenAuth = new URLSearchParams(window.location.search).get("signin");
-  const onLogin = () =>
-    loginWithRedirect({
-      appState: { returnTo: window.location.pathname + window.location.search },
-    });
-
-  if (forceOpenAuth) {
-    onLogin();
-    return null;
-  }
-
-  useEffect(() => {
-    setUserInBrowserPrefs(null);
-  }, []);
-
-  return (
-    <OnboardingModalContainer theme="light">
-      <OnboardingContentWrapper overlay>
-        {isTeamMemberInvite() ? <h1 className="text-2xl font-extrabold">Almost there!</h1> : null}
-        <div className="text-base space-y-4 self-start">
-          {isTeamMemberInvite() ? (
-            <p>In order to join your team, we first need you to sign in.</p>
-          ) : (
-            <>
-              <p className="text-center">
-                Replay captures everything you need for the perfect bug report, all in one link.{" "}
-                <a href="https://replay.io" className="underline pointer-hand">
-                  Learn more
-                </a>
-              </p>
-              <p></p>
-            </>
-          )}
-        </div>
-        <PrimaryLgButton color="blue" onClick={onLogin} className="w-full justify-center">
-          {isTeamMemberInvite() ? "Sign in with Google" : "Login"}
-        </PrimaryLgButton>
-      </OnboardingContentWrapper>
-    </OnboardingModalContainer>
-  );
 }
 
 export default function Account() {
@@ -84,7 +38,7 @@ export default function Account() {
   }
 
   if (!isAuthenticated) {
-    return <WelcomePage />;
+    return <Login />;
   }
 
   return <Library />;
